+++ src/ManagedServices/ManagedServices/Commands/RemoveAzureRmManagedServcicesAssignment.cs
@@ -15,6 +15,7 @@
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.PowerShell.Cmdlets.ManagedServices.Extensions;
 using Microsoft.Azure.PowerShell.Cmdlets.ManagedServices.Models;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using System;
 using System.Management.Automation;
@@ -37,10 +38,20 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.ManagedServices.Commands
         [ScopeCompleter]
         public string Scope { get; set; }
 
+        [CmdletParameterBreakingChange(
+            nameOfParameterChanging: "Id",
+            deprecateByVersion: ManagedServicesUtility.UpcomingVersion,
+            changeInEfectByDate: ManagedServicesUtility.UpcomingVersionReleaseDate,
+            ReplaceMentCmdletParameterName = "Name")]
         [Parameter(ParameterSetName = DefaultParameterSet, Mandatory = true, HelpMessage = "The registration assignment Id.")]
         [ValidateNotNullOrEmpty]
         public string Id { get; set; }
 
+        [CmdletParameterBreakingChange(
+            nameOfParameterChanging: "ResourceId",
+            deprecateByVersion: ManagedServicesUtility.UpcomingVersion,
+            changeInEfectByDate: ManagedServicesUtility.UpcomingVersionReleaseDate,
+            ChangeDescription = ManagedServicesUtility.DeprecatedParameterDescription)]
         [Parameter(ParameterSetName = ByResourceIdParameterSet, Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "The fully qualified rsource id of the registration assignment.")]
         [ValidateNotNullOrEmpty]
         public string ResourceId { get; set; }